+++ master/libspider/spider/platformPThread.h
@@ -40,25 +40,19 @@
 #ifndef PLATFORM_PTHREAD_H
 #define PLATFORM_PTHREAD_H
 
-// semaphore.h includes _ptw32.h that redefines types int64_t and uint64_t on Visual Studio,
-// making compilation error with the IDE's own declaration of said types
-#include <semaphore.h>
+#include "SpiderSemaphore.h"
+
 #include <pthread.h>
+
+#ifdef __APPLE__
+#include "mac_barrier.h"
+#endif
+
 #include <csignal>
 #include <queue>
 #include <map>
 
-#ifdef _MSC_VER
-#ifdef int64_t
-#undef int64_t
-#endif
-
-#ifdef uint64_t
-#undef uint64_t
-#endif
-#endif
-
-#include <platform.h>
+#include "platform.h"
 
 #ifdef PAPI_AVAILABLE
 
@@ -66,11 +60,11 @@
 
 #endif
 
-#include <PThreadSpiderCommunicator.h>
-#include <PThreadLrtCommunicator.h>
-#include <ControlMessageQueue.h>
-#include <NotificationQueue.h>
-#include <SpiderException.h>
+#include "PThreadSpiderCommunicator.h"
+#include "PThreadLrtCommunicator.h"
+#include "ControlMessageQueue.h"
+#include "NotificationQueue.h"
+#include "SpiderException.h"
 
 
 class PlatformPThread : public Platform {